requirements.txt
google-generativeai
python-dotenv
ffmpeg-python
pydantic